requirements.txt
blinker==1.4
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-DebugToolbar==0.10.1
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
nh-currency==1.0.1
requests==2.20.1
urllib3==1.24.1
Werkzeug==0.14.1
